requirements.txt
PyYAML==3.11
argparse==1.2.1
distribute==0.6.24
paho-mqtt==1.1
requests==2.7.0
soco==0.11.1
wsgiref==0.1.2
